backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0

# Static analysis tools
bandit==1.7.6
semgrep==1.54.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# PDF Report Generation
reportlab==4.0.8

# Performance Testing (optional)
locust==2.20.1

# ML dependencies (for future use)
# Uncomment when implementing GNN + LSTM
# torch==2.1.2
# dgl==1.1.3
# transformers==4.36.2
# tree-sitter==0.20.4
